+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.9]
+
+### Changed
+
+- MountFS and MultiFS now accept FS URLS
+- Add openers for AppFS
+
 ## [2.0.8] - 2017-08-13
 
 ### Added
@@ -11,7 +18,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Link info namespace
 - FS.islink method
 - Info.is_link method
-
 
 ## [2.0.7] - 2017-08-06